feat: UART logging, ds18b20 resolution

+++ pid_control.h
@@ -9,7 +9,7 @@
 
 struct pid_control
 {
-    uint8_t set_point;
+    float set_point;
     bool heating_on;
 };
